EntityTable component
