--- packages/browser/src/components/series/table/types.ts
@@ -1,25 +1,0 @@
-import { OnChangeFn, SortingState } from '@tanstack/react-table'
-
-type SeriesTableSorting =
-	| {
-			sorting: SortingState
-			setSorting: OnChangeFn<SortingState>
-	  }
-	| {
-			sorting?: never
-			setSorting?: never
-	  }
-
-type SeriesTableSearch =
-	| {
-			globalFilter: string
-	  }
-	| {
-			globalFilter?: never
-	  }
-
-type StateOptions = SeriesTableSorting & SeriesTableSearch
-
-export type SeriesTableOptions = {
-	enableMultiSort?: boolean
-} & StateOptions
