+++ packages/browser/src/components/table/index.ts
@@ -1,3 +1,8 @@
+export {
+	default as EntityTable,
+	type EntityTableOptions,
+	type EntityTableProps,
+} from './EntityTable'
 export { default as TablePagination, type TablePaginationProps } from './Pagination'
 export { default as SortIcon } from './SortIcon'
 export { getTableModels, default as Table } from './Table'
